// src/services/types/platform/video_recording.ts
export type VideoRecordingApiResponse = {
  _id: string
  __v: number
  video_recording: {
    section1: {
      title1_en: string
      title1_fa: string
      title2_en: string
      title2_fa: string
      text1_en: string
      text1_fa: string
      button1_en: string
      button1_fa: string
    }
    section2_tab1: {
      image1: string
      title1_en: string
      title1_fa: string
      video1: string
    }
    section2_tab2: {
      image1: string
      title1_en: string
      title1_fa: string
      video1: string
    }
    section3: {
      title1_en: string
      title1_fa: string
      text1_en: string
      text1_fa: string
      image1: string
    }
    section4: {
      image1: string
      image2: string
      title1_en: string
      title1_fa: string
      title2_en: string
      title2_fa: string
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
    }
    section5: {
      title1_en: string
      title1_fa: string
      text1_en: string
      text1_fa: string
      image1: string
    }
    section6: {
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
    }
    section7: {
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
      button1_en: string
      button1_fa: string
    }
  }
}